Added responiveness and componetized
